refactor(documents): rebuild the documents sidebar as a right panel
